+++ projects/worldskills-angular-lib/src/lib/common/util/router.util.ts
@@ -1,0 +1,10 @@
+import {NavigationEnd, Router} from '@angular/router';
+import {Observable} from 'rxjs';
+import {filter, map, startWith} from 'rxjs/operators';
+
+export class RouterUtil {
+    static lastNavigationEventWithOrInitialObservable(router: Router, name = null): Observable<any> {
+        return router.events.pipe(filter(event => event instanceof NavigationEnd),
+            startWith(true), map(v => name !== null ? name : v));
+    }
+}